Refactor: many functions moved into their own files
