+++ my-app/src/Components/4 - RelatedComparison/RelatedItems.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { ScrollMenu, VisibilityContext } from 'react-horizontal-scrolling-menu';
+import { ScrollMenu } from 'react-horizontal-scrolling-menu';
 import 'react-horizontal-scrolling-menu/dist/styles.css';
 import axiosAtelier from '../../axiosAtelier.js';
 import Card from './Card.jsx';
 import './hideScrollbar.css';
+import { LeftArrow, RightArrow } from './Arrow.jsx';
 
 export default function App(ogProduct) {
   const [items, setItems] = React.useState([]);
@@ -72,8 +73,6 @@ export default function App(ogProduct) {
     }
   }
 
-  const didMount = React.useRef(false);  //this fixes an issue with waiting for the getProduct call in RelatedComparision to complete.... it sometimes calls twice depending on the speed of the response
-
   React.useEffect(() => {
      getRelatedProducts();
   }, [ogProduct]);
@@ -120,52 +119,3 @@ export default function App(ogProduct) {
     </ScrollMenu>
   );
 }
-
-function Arrow({
-  children,
-  disabled,
-  onClick
-}) {
-  return (
-    <button
-      disabled={disabled}
-      onClick={onClick}
-      style={{
-        cursor: "pointer",
-        display: "flex",
-        flexDirection: "column",
-        justifyContent: "center",
-        right: "1%",
-        opacity: disabled ? "0" : "1",
-        userSelect: "none"
-      }}
-    >
-      {children}
-    </button>
-  );
-}
-
-function LeftArrow() {
-  const { isFirstItemVisible, scrollPrev } =
-    React.useContext(VisibilityContext);
-
-  return (
-    // eslint-disable-next-line react/jsx-no-undef
-    <Arrow disabled={isFirstItemVisible} onClick={() => scrollPrev()}>
-      <a className="btn btn-circle">❮</a>
-    </Arrow>
-  );
-}
-
-function RightArrow() {
-  const { isLastItemVisible, scrollNext } = React.useContext(VisibilityContext);
-
-  return (
-    // eslint-disable-next-line react/jsx-no-undef
-    <Arrow disabled={isLastItemVisible} onClick={() => scrollNext()}>
-      <a className="btn btn-circle">❯</a>
-    </Arrow>
-  );
-}
-
-
